feat: add a select with predefined data | add manual time reasons
