+++ apps/web/lib/features/task/task-filters.tsx
@@ -7,33 +7,26 @@ import {
 	useDailyPlan,
 	useOrganizationTeams,
 	useOutsideClick,
-	useModal,
-	useTeamTasks,
+	useModal
 } from '@app/hooks';
 import { IClassName, ITeamTask } from '@app/interfaces';
 import { clsxm } from '@app/utils';
 import { Transition } from '@headlessui/react';
-import { Button, InputField, SelectItems, Tooltip, VerticalSeparator } from 'lib/components';
+import { Button, InputField, Tooltip, VerticalSeparator } from 'lib/components';
 import { SearchNormalIcon } from 'assets/svg';
 import intersection from 'lodash/intersection';
-import { useCallback, useEffect, useMemo, useState, FormEvent } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 import { TaskUnOrAssignPopover } from './task-assign-popover';
 import { TaskLabelsDropdown, TaskPropertiesDropdown, TaskSizesDropdown, TaskStatusDropdown } from './task-status';
 import { useTranslations } from 'next-intl';
 import { SettingFilterIcon } from 'assets/svg';
 import { DailyPlanFilter } from './daily-plan/daily-plan-filter';
-import { Modal, Divider } from 'lib/components';
-import api from '@app/services/client/axios';
-import { DatePicker } from '@components/ui/DatePicker';
-import { PencilSquareIcon } from '@heroicons/react/20/solid';
-import { FaRegCalendarAlt } from "react-icons/fa";
+import { Divider } from 'lib/components';
+
 import { useDateRange } from '@app/hooks/useDateRange';
 import { TaskDatePickerWithRange } from './task-date-range';
-import { format } from 'date-fns';
-import { HiMiniClock } from "react-icons/hi2";
-import '../../../styles/style.css'
-import { TaskIssueStatus } from './task-issue';
-
+import '../../../styles/style.css';
+import { AddManualTimeModal } from '../manual-time/add-manual-time-modal';
 
 export type ITab = 'worked' | 'assigned' | 'unassigned' | 'dailyplan';
 
@@ -189,9 +182,9 @@ export function useTaskFilter(profile: I_UserProfilePage) {
 					.every((k) => {
 						return k === 'label'
 							? intersection(
-								statusFilters[k],
-								task['tags'].map((item) => item.name)
-							).length === statusFilters[k].length
+									statusFilters[k],
+									task['tags'].map((item) => item.name)
+								).length === statusFilters[k].length
 							: statusFilters[k].includes(task[k]);
 					});
 			});
@@ -276,105 +269,12 @@ export function TaskFilter({ className, hook, profile }: IClassName & Props) {
 function InputFilters({ hook, profile }: Props) {
 	const t = useTranslations();
 	const [loading, setLoading] = useState(false);
-	const { tasks } = useTeamTasks();
-	const { activeTeam } = useOrganizationTeams();
-	const members = activeTeam?.members;
 
-
-	const [date, setDate] = useState<Date>(new Date());
-	const [isBillable, setIsBillable] = useState<boolean>(false);
-	const [startTime, setStartTime] = useState<string>('');
-	const [endTime, setEndTime] = useState<string>('');
-	const [team, setTeam] = useState<string>('');
-	const [task, setTask] = useState<string>('');
-	const [description, setDescription] = useState<string>('');
-	const [reason, setReason] = useState<string>('');
-	const [timeDifference, setTimeDifference] = useState<string>('');
-	const [errorMsg, setError] = useState<string>('');
-	const [loading1, setLoading1] = useState<boolean>(false);
-
-	const { isOpen, openModal, closeModal } = useModal();
-
-	useEffect(() => {
-		const now = new Date();
-		const currentTime = now.toTimeString().slice(0, 5);
-
-		setDate(now);
-		setStartTime(currentTime);
-		setEndTime(currentTime);
-	}, []);
-
-
-	const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-		e.preventDefault();
-
-		const timeObject = {
-			date,
-			isBillable,
-			startTime,
-			endTime,
-			team,
-			task,
-			description,
-			reason,
-			timeDifference
-		};
-
-		if (date && startTime && endTime && team && task) {
-			setLoading1(true);
-			setError('');
-			const postData = async () => {
-				try {
-					const response = await api.post('/add_time', timeObject);
-					if (response.data.message) {
-						setLoading1(false);
-						closeModal();
-					}
-
-				} catch (err) {
-					setError('Failed to post data');
-					setLoading1(false);
-				}
-			};
-
-			postData();
-		} else {
-			setError(`Please complete all required fields with a ${"*"}`)
-		}
-	};
-
-	const calculateTimeDifference = () => {
-
-		const [startHours, startMinutes] = startTime.split(':').map(Number);
-		const [endHours, endMinutes] = endTime.split(':').map(Number);
-
-		const startTotalMinutes = startHours * 60 + startMinutes;
-		const endTotalMinutes = endHours * 60 + endMinutes;
-
-		const diffMinutes = endTotalMinutes - startTotalMinutes;
-		if (diffMinutes < 0) {
-			return;
-		}
-
-		const hours = Math.floor(diffMinutes / 60);
-		const minutes = diffMinutes % 60;
-		setTimeDifference(`${String(hours).padStart(2, '0')}h ${String(minutes).padStart(2, '0')}m`);
-	};
-
-
-	useEffect(() => {
-		calculateTimeDifference();
-	}, [endTime, startTime]);
-
-	useEffect(() => {
-		if (task == '') {
-			setTask(tasks[0]?.id);
-		}
-		if (team == '') {
-			members && setTeam(members[0].id);
-		}
-
-	}, [tasks, members]);
+	const {
+		isOpen: isManualTimeModalOpen,
+		openModal: openManualTimeModal,
+		closeModal: closeManualTimeModal
+	} = useModal();
 
 	const osSpecificAssignTaskTooltipLabel = 'A';
 
@@ -406,13 +306,13 @@ function InputFilters({ hook, profile }: Props) {
 				<span>{t('common.FILTER')}</span>
 			</button>
 			<Button
-				onClick={() => openModal()}
+				onClick={() => openManualTimeModal()}
 				className={clsxm(
 					'bg-primary text-white dark:border-from-regal-rose dark:border-to-regal-blue dark:text-transparent dark:bg-clip-text dark:bg-gradient-to-tl  dark:from-regal-rose dark:to-regal-blue h-full px-4 py-3 rounded-xl text-base flex items-center space-x-1 dark:border-gradient-dark dark:border-regal-rose dark:border',
 					'min-w-[8.25rem] h-[2.75rem]'
 				)}
 			>
-				<span className='text-xl'>+</span>
+				<span className="text-xl">+</span>
 				Add time
 			</Button>
 
@@ -441,163 +341,8 @@ function InputFilters({ hook, profile }: Props) {
 					</Button>
 				</Tooltip>
 			</TaskUnOrAssignPopover>
-
-			<Modal
-				isOpen={isOpen}
-				closeModal={closeModal}
-				title={'Add time'}
-				className="bg-light--theme-light dark:bg-dark--theme-light py-5 rounded-xl w-full md:min-w-[25vw] md:max-w-fit h-[auto] justify-start shadow-xl"
-				titleClass="text-[16px] font-bold"
-			>
-				{/* <Divider className="mt-4" /> */}
-				<form onSubmit={handleSubmit} className="text-[13px] w-[90%]">
-					<div className="mb-4">
-						<label className="block text-gray-500 mb-1">Date<span className="text-[#de5505e1] ml-1">*</span></label>
-						<div className="w-full p-2 border border-slate-100 dark:border-slate-600 rounded-[10px]">
-							<DatePicker
-								buttonVariant={'link'}
-								className='dark:bg-dark--theme-light'
-								buttonClassName={'decoration-transparent h-[0.875rem] w-full flex items-center'}
-								customInput={
-									<div
-										className={clsxm(
-											'not-italic cursor-pointer font-semibold text-[0.625rem] 3xl:text-xs w-full',
-											'leading-[140%] tracking-[-0.02em] text-[#282048] dark:text-white w-full'
-										)}
-									>
-										{
-											date ?
-												<div className="flex items-center w-full">
-													<FaRegCalendarAlt size={20} fill={"#3826A6"} className="mx-[20px] text-[#3826A6]" />
-													{format(date, "PPP")}
-												</div>
-												: (
-													<PencilSquareIcon className="dark:text-white text-dark w-4 h-4" />
-												)}
-									</div>
-								}
-								selected={date}
-								onSelect={(value) => {
-									value && setDate(value);
-								}}
-								mode={'single'}
-							/>
-						</div>
-					</div>
-
-					<div className="mb-4 flex items-center">
-						<label className="block text-gray-500 mr-2">Billable</label>
-						<div
-							className={`w-12 h-6 flex items-center bg-[#6c57f4b7] rounded-full p-1 cursor-pointer `}
-							onClick={() => setIsBillable(!isBillable)}
-							style={isBillable ? { background: 'linear-gradient(to right, #9d91efb7, #8a7bedb7)' } : { background: '#6c57f4b7' }}
-						>
-							<div
-								className={` bg-[#3826A6] w-4 h-4 rounded-full shadow-md transform transition-transform ${isBillable ? 'translate-x-6' : 'translate-x-0'}`}
-							/>
-						</div>
-					</div>
-					<div className='flex items-center'>
-						<div className="mb-4 w-[48%] mr-[4%]">
-							<label className="block text-gray-500 mb-1">Start time<span className="text-[#de5505e1] ml-1">*</span></label>
-							<input
-								type="time"
-								value={startTime}
-								onChange={(e) => setStartTime(e.target.value)}
-								className="w-full p-2 border text-[13px] font-bold border-slate-100 dark:border-slate-600 rounded-[10px] dark:bg-dark--theme-light"
-								required
-							/>
-						</div>
-
-						<div className="mb-4 w-[48%]"><label className="block text-gray-500 mb-1">End time<span className="text-[#de5505e1] ml-1">*</span></label>
-
-							<input
-								type="time"
-								value={endTime}
-								onChange={(e) => setEndTime(e.target.value)}
-								className="w-full p-2 border text-[13px] font-bold border-slate-100 dark:border-slate-600 rounded-[10px] dark:bg-dark--theme-light"
-								required
-							/>
-						</div>
-					</div>
-
-					<div className="mb-4 flex items-center">
-						<label className="block text-gray-500 mb-1">Added hours: </label>
-						<div
-							className="ml-[10px] p-1 flex items-center font-bold dark:border-regal-rose border rounded-[10px] pr-3"
-						// style={{ background: 'linear- (to right, #0085c8a4, #410a504e )' }}
-						>
-							<div className='mr-[10px] bg-gradient-to-tl from-[#3826A6] dark:from-regal-rose to-regal-blue rounded-full '>
-								<HiMiniClock size={20} className="bg-gradient-to-tl from-regal-rose to-regal-blue rounded-full text-[#8a7bedb7] dark:text-to-regal-rose" />
-							</div>
-							{timeDifference}
-						</div>
-					</div>
-
-					<div className="mb-4">
-						<label className="block text-gray-500 mb-1">Team<span className="text-[#de5505e1] ml-1">*</span></label>
-						<SelectItems items={members!}
-							onValueChange={(value) => setTeam(value)}
-							itemId={(member) => (member.id)}
-							itemToString={(member: any) => member?.employee?.user?.firstName}
-							triggerClassName='border-slate-100 dark:border-slate-600'
-						/>
-					</div>
-
-					<div className="mb-4">
-						<label className="block text-gray-500 mb-1">Task<span className="text-[#de5505e1] ml-1">*</span></label>
-						<SelectItems
-							items={tasks}
-							itemToString={(task) => task.title}
-							itemId={(task) => task.id}
-							onValueChange={(value) => setTask(value)}
-							triggerClassName='border-slate-100 dark:border-slate-600'
-							renderItem={(item, onClick) => {
-								return <button
-									style={{ textOverflow: 'ellipsis', whiteSpace: 'nowrap', overflow: 'hidden' }}
-									key={item.id} onClick={onClick} className={clsxm("truncate hover:cursor-pointer hover:bg-slate-50 w-full text-[13px] hover:rounded-lg p-1 hover:font-bold dark:text-white dark:hover:bg-primary flex items-center space-x-1 justify-start")}
-								>
-									<TaskIssueStatus showIssueLabels={false} className={'truncate hover:cursor-pointer h-5'} task={item} />
-									<span
-										style={{ textOverflow: 'ellipsis', whiteSpace: 'nowrap', overflow: 'hidden' }}
-									>
-										{item.title}
-									</span>
-								</button>
-							}}
-						/>
-					</div>
-
-					<div className="mb-4">
-						<label className="block text-gray-500 mb-1">Description</label>
-						<textarea
-							value={description}
-							placeholder="What worked on?"
-							onChange={(e) => setDescription(e.target.value)}
-							className="w-full p-2 border border-slate-100 dark:border-slate-600 rounded-[10px] dark:bg-dark--theme-light"
-						/>
-					</div>
-
-					<div className="mb-4">
-						<label className="block text-gray-500 mb-1">Reason</label>
-						<textarea
-							value={reason}
-							onChange={(e) => setReason(e.target.value)}
-							className="w-full p-2 border border-slate-100 dark:border-slate-600 rounded-[10px] dark:bg-dark--theme-light"
-						/>
-					</div>
-
-					<div className="flex justify-between items-center w-full">
-						<Button variant='outline' type="button" className="dark:text-primary border-slate-100 dark:border-slate-600 font-bold p-[12px] h-11 rounded-[10px] border-[0.5px] dark:bg-dark--theme-light">View timesheet</Button>
-						<Button loading={loading1} disabled={loading1} type="submit" className="bg-[#3826A6] font-bold  flex items-center text-white p-[12px] rounded-[10px]">
-							Add time
-						</Button>
-					</div>
-					<div className="m-4 text-[#ff6a00de]">{errorMsg}</div>
-
-				</form>
-			</Modal>
-		</div >
+			<AddManualTimeModal closeModal={closeManualTimeModal} isOpen={isManualTimeModalOpen} />
+		</div>
 	);
 }
 
@@ -687,7 +432,12 @@ export function TaskStatusFilter({ hook, employeeId }: { hook: I_TaskFilter; emp
 
 				{hook.tab === 'dailyplan' && <DailyPlanFilter employeeId={employeeId} />}
 				{['Future Tasks', 'Past Tasks', 'All Tasks'].includes(dailyPlanTab) && (
-					<TaskDatePickerWithRange data={data} date={date} onSelect={(range) => setDate(range)} label="Planned date" />
+					<TaskDatePickerWithRange
+						data={data}
+						date={date}
+						onSelect={(range) => setDate(range)}
+						label="Planned date"
+					/>
 				)}
 				<VerticalSeparator />
